feat(otel): Upgrade @opentelemetry/semantic-conventions to 1.26.0
In 1.26.0 otel-js has updated the deprecations for the attributes based
on the new changes to semantic conventions.

They also changed the name of every export, for example:
`SEMATTRS_HTTP_ROUTE` -> `ATTR_HTTP_ROUTE`.

In addition some of these exports now live in
`@opentelemetry/semantic-conventions/incubating` which means we have to
change the overall export.